+++ example/web_crawler.dart
@@ -1,7 +1,6 @@
 import 'package:dart_toolkit/cli.dart';
-import 'package:dart_toolkit/core.dart';
 import 'package:dart_toolkit/http.dart';
-import 'package:dart_toolkit/util.dart';
+import 'package:dart_toolkit/core.dart';
 
 typedef Story = ({String title, Uri link});