+++ lib/dart_toolkit.dart
@@ -1,14 +1,12 @@
 /// # Dart Toolkit
 library;
 
-export 'archive.dart';
 export 'async.dart';
 export 'cli.dart';
 export 'collection.dart';
 export 'core.dart';
 export 'formats.dart';
 export 'fs.dart';
-export 'hash.dart';
+export 'crypto.dart';
 export 'http.dart';
 export 'process.dart';
-export 'util.dart';